feat: add facebook video embed support
